+++ visualizer/package.json
@@ -22,6 +22,7 @@
     "vite": "^5.4.9"
   },
   "dependencies": {
-    "@coderline/alphatab": "^1.3.1"
+    "@coderline/alphatab": "^1.3.1",
+    "svelte-spa-router": "^4.0.1"
   }
 }
