+++ visualizer/src/routes/Home.svelte
@@ -1,0 +1,6 @@
+<div class="flex flex-col items-center justify-center min-h-screen bg-gray-100">
+  <h1 class="text-4xl font-bold text-gray-800 mb-4">Welcome to Our App!</h1>
+  <p class="text-lg text-gray-600 mb-6">
+    Explore and navigate through the app to discover more.
+  </p>
+</div>
